Rework: Price transport
